fix: change folder name
